MacOS compile fixes

+++ src/io/socket/tcpsocket.cpp
@@ -236,7 +236,7 @@ bool TcpSocket::getMessage(std::string& message) {
     }
     std::lock_guard<std::mutex> inputLock(_inputQueueMutex);
     message = std::string(_inputQueue.begin(), _inputQueue.begin() + delimiterIndex);
-    if (_inputQueue.size() >= delimiterIndex + 1) {
+    if (static_cast<int>(_inputQueue.size()) >= delimiterIndex + 1) {
         _inputQueue.erase(_inputQueue.begin(), _inputQueue.begin() + delimiterIndex + 1);
     }
     return true;
